+++ benchmark/benchmark_curand_generate.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2021 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -205,7 +205,7 @@ const std::vector<std::string> all_engines = {
     "philox",
     "sobol32",
     // "scrambled_sobol32",
-    // "sobol64",
+    "sobol64",
     // "scrambled_sobol64",
 };