+++ library/src/rng/mrg32k3a.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2021 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -38,7 +38,7 @@ namespace detail {
     typedef ::rocrand_device::mrg32k3a_engine mrg32k3a_device_engine;
 
     __global__
-    __launch_bounds__(ROCRAND_DEFAULT_MAX_BLOCK_SIZE, ROCRAND_DEFAULT_MIN_WARPS_PER_EU)
+    __launch_bounds__(ROCRAND_DEFAULT_MAX_BLOCK_SIZE)
     void init_engines_kernel(mrg32k3a_device_engine * engines,
                              unsigned long long seed,
                              unsigned long long offset)
@@ -49,7 +49,7 @@ namespace detail {
 
     template<class T, class Distribution>
     __global__
-    __launch_bounds__(ROCRAND_DEFAULT_MAX_BLOCK_SIZE, ROCRAND_DEFAULT_MIN_WARPS_PER_EU)
+    __launch_bounds__(ROCRAND_DEFAULT_MAX_BLOCK_SIZE)
     void generate_kernel(mrg32k3a_device_engine * engines,
                          T * data, const size_t n,
                          Distribution distribution)